blindly tried to fix a 32bit memory issue by adding random uint64 casts

+++ draw.h
@@ -9,6 +9,6 @@ bool loadImagePartBmp(FILE* fh, int startx, int starty, int width, int height);
 void setPixelBmp(size_t x, size_t y, uint8_t color, float fsub);
 void blendPixelBmp(size_t x, size_t y, uint8_t color, float fsub);
 bool saveImagePartBmp(FILE* fh);
-size_t calcImageSizeBmp(int mapChunksX, int mapChunksZ, size_t mapHeight, int &pixelsX, int &pixelsY, bool tight = false);
+uint64_t calcImageSizeBmp(int mapChunksX, int mapChunksZ, size_t mapHeight, int &pixelsX, int &pixelsY, bool tight = false);
 
 #endif
